+++ core/processor/ProcessorParseTimestampNative.cpp
@@ -20,6 +20,7 @@
 #include "common/Constants.h"
 #include "common/LogtailCommonFlags.h"
 #include "plugin/ProcessorInstance.h"
+#include <algorithm>
 
 
 namespace logtail {
@@ -31,11 +32,16 @@ bool ProcessorParseTimestampNative::Init(const ComponentConfig& componentConfig)
     mLegacyPreciseTimestampConfig.enabled = config.mAdvancedConfig.mEnablePreciseTimestamp;
     mLegacyPreciseTimestampConfig.key = config.mAdvancedConfig.mPreciseTimestampKey;
     mLegacyPreciseTimestampConfig.unit = config.mAdvancedConfig.mPreciseTimestampUnit;
-    mLogTimeZoneOffsetSecond = config.mLogTimeZoneOffsetSecond;
+    mLogTimeZoneOffsetSecond = config.mTimeZoneAdjust ? config.mLogTimeZoneOffsetSecond  - GetLocalTimeZoneOffsetSecond() : 0;
 
     mParseTimeFailures = &(GetContext().GetProcessProfile().parseTimeFailures);
     mHistoryFailures = &(GetContext().GetProcessProfile().historyFailures);
     SetMetricsRecordRef(Name(), componentConfig.GetId());
+    mProcParseInSizeBytes = GetMetricsRecordRef().CreateCounter(METRIC_PROC_PARSE_IN_SIZE_BYTES);
+    mProcParseOutSizeBytes = GetMetricsRecordRef().CreateCounter(METRIC_PROC_PARSE_OUT_SIZE_BYTES);
+    mProcDiscardRecordsTotal = GetMetricsRecordRef().CreateCounter(METRIC_PROC_DISCARD_RECORDS_TOTAL);
+    mProcParseErrorTotal = GetMetricsRecordRef().CreateCounter(METRIC_PROC_PARSE_ERROR_TOTAL);
+    mProcHistoryFailureTotal = GetMetricsRecordRef().CreateCounter(METRIC_PROC_HISTORY_FAILURE_TOTAL);
     return true;
 }
 
@@ -47,8 +53,9 @@ void ProcessorParseTimestampNative::Process(PipelineEventGroup& logGroup) {
     StringView timeStrCache;
     EventsContainer& events = logGroup.MutableEvents();
     // works good normally. poor performance if most data need to be discarded.
+    LogtailTime logTime = {0, 0};
     for (auto it = events.begin(); it != events.end();) {
-        if (ProcessorParseTimestampNative::ProcessEvent(logPath, *it, timeStrCache)) {
+        if (ProcessorParseTimestampNative::ProcessEvent(logPath, *it, logTime, timeStrCache)) {
             ++it;
         } else {
             it = events.erase(it);
@@ -61,7 +68,7 @@ bool ProcessorParseTimestampNative::IsSupportedEvent(const PipelineEventPtr& e) 
     return e.Is<LogEvent>();
 }
 
-bool ProcessorParseTimestampNative::ProcessEvent(StringView logPath, PipelineEventPtr& e, StringView& timeStrCache) {
+bool ProcessorParseTimestampNative::ProcessEvent(StringView logPath, PipelineEventPtr& e, LogtailTime& logTime, StringView& timeStrCache) {
     if (!IsSupportedEvent(e)) {
         return true;
     }
@@ -70,7 +77,7 @@ bool ProcessorParseTimestampNative::ProcessEvent(StringView logPath, PipelineEve
         return true;
     }
     const StringView& timeStr = sourceEvent.GetContent(mTimeKey);
-    LogtailTime logTime = {0, 0};
+    mProcParseInSizeBytes->Add(timeStr.size());
     uint64_t preciseTimestamp = 0;
     bool parseSuccess = ParseLogTime(timeStr, logPath, logTime, preciseTimestamp, timeStrCache);
     if (!parseSuccess) {
@@ -79,7 +86,7 @@ bool ProcessorParseTimestampNative::ProcessEvent(StringView logPath, PipelineEve
     if (logTime.tv_sec <= 0
         || (BOOL_FLAG(ilogtail_discard_old_data)
             // Adjust time(NULL) from local timezone to target timezone
-            && (time(NULL) - logTime.tv_sec) > INT32_FLAG(ilogtail_discard_interval))) {
+            && (time(NULL) - mLogTimeZoneOffsetSecond - logTime.tv_sec) > INT32_FLAG(ilogtail_discard_interval))) {
         if (AppConfig::GetInstance()->IsLogParseAlarmValid()) {
             if (LogtailAlarm::GetInstance()->IsLowLevelAlarmValid()) {
                 LOG_WARNING(sLogger,
@@ -96,18 +103,17 @@ bool ProcessorParseTimestampNative::ProcessEvent(StringView logPath, PipelineEve
                                                    GetContext().GetRegion());
         }
         ++(*mHistoryFailures);
+        mProcHistoryFailureTotal->Add(1);
+        mProcDiscardRecordsTotal->Add(1);
         return false;
     }
-    sourceEvent.SetTimestamp(logTime.tv_sec);
+    sourceEvent.SetTimestamp(logTime.tv_sec, logTime.tv_nsec);
     if (mLegacyPreciseTimestampConfig.enabled) {
-        int preciseTimestampLen = 0;
-        do {
-            ++preciseTimestampLen;
-        } while ((preciseTimestamp /= 10) > 0);
-        StringBuffer sb = sourceEvent.GetSourceBuffer()->AllocateStringBuffer(preciseTimestampLen + 1);
-        snprintf(sb.data, sb.capacity, "%lu", preciseTimestamp);
+        StringBuffer sb = sourceEvent.GetSourceBuffer()->AllocateStringBuffer(20);
+        sb.size = std::min(20, snprintf(sb.data, sb.capacity, "%lu", preciseTimestamp));
         sourceEvent.SetContentNoCopy(mLegacyPreciseTimestampConfig.key, StringView(sb.data, sb.size));
     }
+    mProcParseOutSizeBytes->Add(sizeof(logTime.tv_sec) + sizeof(logTime.tv_nsec));
     return true;
 }
 
@@ -126,7 +132,8 @@ bool ProcessorParseTimestampNative::ParseLogTime(const StringView& curTimeStr, /
     int nanosecondLength = -1;
     const char* strptimeResult = NULL;
     if ((!haveNanosecond || endWithNanosecond) && IsPrefixString(curTimeStr, timeStrCache)) {
-        if (endWithNanosecond) {
+        bool isTimestampNanosecond = (mTimeFormat == "%s") && (curTimeStr.length() > timeStrCache.length());
+        if (endWithNanosecond || isTimestampNanosecond) {
             strptimeResult = Strptime(curTimeStr.data() + timeStrCache.length(), "%f", &logTime, nanosecondLength);
         } else {
             strptimeResult = curTimeStr.data() + timeStrCache.length();
@@ -134,8 +141,10 @@ bool ProcessorParseTimestampNative::ParseLogTime(const StringView& curTimeStr, /
         }
     } else {
         strptimeResult = Strptime(curTimeStr.data(), mTimeFormat.c_str(), &logTime, nanosecondLength, mSpecifiedYear);
-        timeStrCache = curTimeStr.substr(0, curTimeStr.length() - nanosecondLength);
-        logTime.tv_sec = logTime.tv_sec - mLogTimeZoneOffsetSecond;
+        if (NULL != strptimeResult) {
+            timeStrCache = curTimeStr.substr(0, curTimeStr.length() - nanosecondLength);
+            logTime.tv_sec = logTime.tv_sec - mLogTimeZoneOffsetSecond;
+        }
     }
     if (NULL == strptimeResult) {
         if (AppConfig::GetInstance()->IsLogParseAlarmValid()) {
@@ -151,6 +160,7 @@ bool ProcessorParseTimestampNative::ParseLogTime(const StringView& curTimeStr, /
                                                    GetContext().GetRegion());
         }
 
+        mProcParseErrorTotal->Add(1);
         ++(*mParseTimeFailures);
         return false;
     }